Initial wrapper API implementation; not for submit
Patch 1 - initial patch
Patch 2 - VbInit() clears recovery request
Patch 3 - type cleanup found in H2C build
Patch 4 - (rebase)
Patch 5 - Fix copyright dates; cleanup host calls to wrapper API

BUG=none
TEST=make && make runtests; compiles and runs on H2C

+++ firmware/lib/cryptolib/include/sha.h
@@ -107,7 +107,7 @@ typedef struct DigestContext {
 
 /* Initialize a digest context for use with signature algorithm [algorithm]. */
 void DigestInit(DigestContext* ctx, int sig_algorithm);
-void DigestUpdate(DigestContext* ctx, const uint8_t* data, uint64_t len);
+void DigestUpdate(DigestContext* ctx, const uint8_t* data, uint32_t len);
 
 /* Caller owns the returned digest and must free it. */
 uint8_t* DigestFinal(DigestContext* ctx);
